Cleanup and DIARY.md updates
