--- ElasticsearchMemoryStorage/ElasticHelper.cs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Free Mind Labs, Inc. All rights reserved.
-
-namespace FreeMindLabs.KernelMemory.Elasticsearch;
-
-/// <summary>
-/// A quick helper class for Elasticsearch.
-/// </summary>
-public static class ElasticHelper
-{
-
-}
