+++ UnitTests/Serverless/ServerlessTest.cs
@@ -66,4 +66,40 @@ public class ServerlessTest
             this._output.WriteLine($"  - {x.SourceName}  - {x.Link} [{x.Partitions.First().LastUpdate:D}]");
         }
     }
+
+    [Fact]
+    public void AllSequencesOfConfigurationsWork(string[] args)
+    {
+        // Concatenate our 'WithElasticsearch()' after 'WithOpenAIDefaults()' from the core nuget
+        var test1 = new KernelMemoryBuilder()
+            .WithOpenAIDefaults("api key")
+            .WithElasticsearch(
+                endpoint: "https://localhost:9200",
+                userName: "elastic",
+                password: "changeme",
+                certificateFingerPrint: "1234567890"
+            )
+            .Build();
+
+        // Concatenate our 'WithElasticsearch()' before 'WithOpenAIDefaults()' from the core nuget
+        var test2 = new KernelMemoryBuilder()
+            .WithElasticsearch(endpoint: "https://localhost:9200",
+                userName: "elastic",
+                password: "changeme",
+                certificateFingerPrint: "1234567890")
+            .WithOpenAIDefaults("api key")
+            .Build();
+
+        // Concatenate our 'WithElasticsearch()' before and after KM builder extension methods from the core nuget
+        var test3 = new KernelMemoryBuilder()
+            .WithSimpleFileStorage()
+            .WithElasticsearch(endpoint: "https://localhost:9200",
+                userName: "elastic",
+                password: "changeme",
+                certificateFingerPrint: "1234567890")
+            .WithOpenAIDefaults("api key")
+            .Build();
+
+        Console.WriteLine("Test complete");
+    }
 }
